Float.

+++ double-conversion/bignum-dtoa.h
@@ -74,7 +74,7 @@ enum BignumDtoaMode {
 //   Halfway cases are again rounded up.
 // 'BignumDtoa' expects the given buffer to be big enough to hold all digits
 // and a terminating null-character.
-void BignumDtoa(double v, BignumDtoaMode mode, int requested_digits,
+void BignumDtoa(float v, BignumDtoaMode mode, int requested_digits,
                 Vector<char> buffer, int* length, int* point);
 
 }  // namespace double_conversion
